+++ bindings/nostr-sdk-ffi/src/protocol/nips/nip47.rs
@@ -924,13 +924,16 @@ impl NostrWalletConnectURI {
     #[uniffi::constructor]
     pub fn new(
         public_key: &PublicKey,
-        relay_url: &str,
+        relays: Vec<String>,
         random_secret_key: &SecretKey,
         lud16: Option<String>,
     ) -> Result<Self> {
         Ok(nip47::NostrWalletConnectURI::new(
             **public_key,
-            RelayUrl::parse(relay_url)?,
+            relays
+                .into_iter()
+                .map(|r| RelayUrl::parse(&r))
+                .collect::<Result<Vec<_>, _>>()?,
             random_secret_key.deref().clone(),
             lud16,
         )
@@ -947,9 +950,9 @@ impl NostrWalletConnectURI {
         Arc::new(self.inner.public_key.into())
     }
 
-    /// URL of the relay of choice where the `App` is connected and the `Signer` must send and listen for messages.
-    pub fn relay_url(&self) -> String {
-        self.inner.relay_url.to_string()
+    /// URLs of the relays of choice where the `App` is connected and the `Signer` must send and listen for messages.
+    pub fn relays(&self) -> Vec<String> {
+        self.inner.relays.iter().map(|r| r.to_string()).collect()
     }
 
     /// 32-byte randomly generated hex encoded string